Update test server for 2.0
